+++ src/board/zobrist.rs
@@ -4,7 +4,7 @@ use rand::{Rng, SeedableRng};
 use crate::defs::{NrOf, Piece, Side, Sides, Square, EMPTY};
 
 const ALL_SQUARES: usize = NrOf::SQUARES;
-const ALL_PIECES: usize = NrOf::PIECE_TYPES as usize;
+const ALL_PIECES: usize = NrOf::PIECE_TYPES;
 const ALL_SIDES: usize = Sides::BOTH;
 const ALL_CASTLING_PERMISSIONS: usize = NrOf::CASTLING_PERMISSIONS as usize;
